html/semantics/forms/the-label-element/label-attributes.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: The label element</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<form id="fm">
  <label id="lbl0" for="test0"></label>
  <b id="test0"></b>

  <label id="lbl1" for="test1"></label>
  <input id="test1" class="test1">
  <label id="lbl2" for="test1"></label>

  <label id="lbl3">
    <a id="test2"></a>
    <div><input id="input1" class="textbox"></div>
    <input id="input2">
  </label>
</form>

<label id="lbl4"></label>

<label id="lbl5" for="input4">
  <b id="test2"></b>
  <div><input id="input3"></div>
</label>
<input id="input4">

<script>

  test(function () {
    assert_not_equals(document.getElementById("lbl0").control, document.getElementById("test0"),
                      "An element that's not a labelable element can't be a label element's labeled control.");
    assert_equals(document.getElementById("lbl0").control, null,
                  "A label element whose 'for' attribute doesn't reference any labelable element shouldn't have any labeled control.");
  }, "A label element with a 'for' attribute should only be associated with a labelable element.");

  test(function () {
    var labels = document.querySelector(".test1").labels;
    assert_true(labels instanceof NodeList,
                "A form control's 'labels' property should be an instance of a NodeList.");
    assert_equals(labels.length, 2,
                  "The number of labels associated with a form control should be the number of label elements for which it is a labeled control.");
    assert_array_equals(labels, [document.getElementById("lbl1"), document.getElementById("lbl2")],
                        "The labels for a form control should be returned in tree order.");

    var newLabel = document.createElement("label");
    newLabel.htmlFor = "test1";
    document.getElementById("fm").insertBefore(newLabel, document.getElementById("lbl0"));
    assert_array_equals(document.querySelector(".test1").labels, [newLabel, document.getElementById("lbl1"), document.getElementById("lbl2")],
                        "The labels for a form control should be returned in tree order.");
  }, "A form control can have multiple labels.");

  test(function () {
    var labels = document.getElementById("input2").labels;
    assert_true(labels instanceof NodeList, "A form control's 'labels' property should be an instance of a NodeList.");
    assert_equals(labels.length, 0, "The number of labels should be 0 if the associated form control isn't referenced by any <label>.");
  }, "A form control has no label 1");

  test(function () {
    assert_equals(document.getElementById("lbl5").control, document.getElementById("input4"),
                  "The labeled cotrol should be associated with the control whose ID is equal to the value of the 'for' attribute.");
    assert_equals(document.getElementById("input3").labels.length, 0,
                  "The number of labels should be 0 if the form control has an ancestor label elment that the for attribute points to another control.");

    document.getElementById("input3").setAttribute("for", "");
    assert_equals(document.getElementById("input3").labels.length, 0,
                  "The number of labels should be 0 if the form control has an ancestor label elment that the for attribute is an empty string.");
  }, "A form control has no label 2");

  test(function () {
    document.getElementById("lbl5").setAttribute("for", "test2");
    assert_equals(document.getElementById("lbl5").control, null,
                  "The label's 'control' property should return null if its 'for' attribute doesn't point to a labelable element.");

    document.getElementById("lbl5").setAttribute("for", "test3");;
    assert_equals(document.getElementById("lbl5").control, null,
                  "The label's 'control' property should return null if its 'for' attribute points to an inexistent element.");
  }, "A form control has no label 3");

  test(function () {
    assert_equals(document.getElementById("lbl3").control, document.querySelector(".textbox"),
                  "The first labelable descendant of a label element should be its labeled control.");
  }, "The labeled control for a label element that has no 'for' attribute is the first labelable element which is a descendant of that label element.");

  test(function () {
    assert_equals(document.getElementById("lbl0").form, document.getElementById("fm"),
                  "The 'form' property for a label with a form owner should return the form owner.");
  }, "A label's form attribute should return its form owner.");

  test(function () {
    assert_equals(document.getElementById("lbl4").form, null,
                  "The 'form' property for a label without a form owner should return null.");
  }, "Check that the labels property of a form control with no label returns a zero-length NodeList.");

</script>
